refactor: split docs markdown core helpers
