+++ crates/ox_content_docs/src/markdown/stats.rs
@@ -1,0 +1,156 @@
+use super::{
+    markdown_html, markdown_pure, MarkdownDisplayFormat, MarkdownDocsOptions, MarkdownRenderStyle,
+    DOC_KIND_ORDER,
+};
+use crate::model::{ApiDocEntry, ApiDocModule};
+
+#[derive(Debug, Clone, Default)]
+pub(super) struct EntryStats {
+    pub(in crate::markdown) entries: usize,
+    pub(in crate::markdown) by_kind: [usize; DOC_KIND_ORDER.len()],
+    pub(in crate::markdown) members: usize,
+    pub(in crate::markdown) params: usize,
+    pub(in crate::markdown) returns: usize,
+    pub(in crate::markdown) examples: usize,
+    pub(in crate::markdown) deprecated: usize,
+}
+
+pub(super) fn summarize_entries<'a>(
+    entries: impl IntoIterator<Item = &'a ApiDocEntry>,
+) -> EntryStats {
+    let mut stats = EntryStats::default();
+
+    for entry in entries {
+        stats.entries += 1;
+        if let Some(index) = doc_kind_index(&entry.kind) {
+            stats.by_kind[index] += 1;
+        }
+        stats.members += entry.members.len();
+        stats.params += entry.params.len();
+        stats.returns += usize::from(entry.returns.is_some());
+        stats.examples += entry.examples.len();
+        stats.deprecated += usize::from(entry.tags.iter().any(|tag| tag.tag == "deprecated"));
+    }
+
+    stats
+}
+
+pub(super) fn summarize_module(module: &ApiDocModule) -> EntryStats {
+    let mut stats = summarize_entries(&module.entries);
+    stats.examples += module.examples.len();
+    stats
+}
+
+pub(super) fn summarize_docs(docs: &[ApiDocModule]) -> EntryStats {
+    let mut stats = summarize_entries(docs.iter().flat_map(|doc| doc.entries.iter()));
+    stats.examples += docs.iter().map(|doc| doc.examples.len()).sum::<usize>();
+    stats
+}
+
+fn doc_kind_index(kind: &str) -> Option<usize> {
+    match kind {
+        "function" => Some(0),
+        "class" => Some(1),
+        "interface" => Some(2),
+        "type" => Some(3),
+        "enum" => Some(4),
+        "variable" => Some(5),
+        "module" => Some(6),
+        _ => None,
+    }
+}
+
+pub(super) fn doc_kind_plural(kind: &str) -> &'static str {
+    match kind {
+        "function" => "functions",
+        "class" => "classes",
+        "interface" => "interfaces",
+        "type" => "types",
+        "enum" => "enumerations",
+        "variable" => "variables",
+        "module" => "modules",
+        _ => "symbols",
+    }
+}
+
+/// Whether a member table should keep the `Kind` column. Named member groups
+/// (`Properties`, `Methods`, `Constructors`, `Enum Members`, `Static …`) state the
+/// kind in their heading, so the column is redundant and dropped to match TypeDoc.
+/// Only the generic `Members` fallback (mixed kinds) keeps it. Shared by both
+/// renderers via `super::member_table_includes_kind`.
+pub(super) fn member_table_includes_kind(group_title: &str) -> bool {
+    group_title == "Members"
+}
+
+/// Renders the per-page stats summary in the configured render style.
+fn render_stats_summary(
+    options: &MarkdownDocsOptions,
+    stats: &EntryStats,
+    module_count: Option<usize>,
+) -> String {
+    match options.render_style {
+        MarkdownRenderStyle::Html => markdown_html::render_stats_html(stats, module_count),
+        MarkdownRenderStyle::Markdown => markdown_pure::render_stats_markdown(stats, module_count),
+    }
+}
+
+/// Appends the stats summary plus its trailing blank line, unless stats are
+/// disabled via `options.render_stats`. Centralizing the gate keeps every
+/// index/overview generator from leaving a stray blank line when stats are
+/// omitted.
+pub(super) fn push_stats(
+    markdown: &mut String,
+    options: &MarkdownDocsOptions,
+    stats: &EntryStats,
+    module_count: Option<usize>,
+) {
+    if !options.render_stats {
+        return;
+    }
+    markdown.push_str(&render_stats_summary(options, stats, module_count));
+    markdown.push_str("\n\n");
+}
+
+/// Appends the generated-by attribution plus its trailing blank line, unless it
+/// is disabled via `options.render_generated_by`.
+pub(super) fn push_generated_by(markdown: &mut String, options: &MarkdownDocsOptions) {
+    if !options.render_generated_by {
+        return;
+    }
+    markdown.push_str("Generated by [Ox Content](https://github.com/ubugeeei-prod/ox-content)\n\n");
+}
+
+fn effective_display_format(
+    options: &MarkdownDocsOptions,
+    format: MarkdownDisplayFormat,
+) -> MarkdownDisplayFormat {
+    match (options.render_style, format) {
+        (MarkdownRenderStyle::Markdown, MarkdownDisplayFormat::None) => MarkdownDisplayFormat::List,
+        (_, format) => format,
+    }
+}
+
+pub(super) fn effective_index_format(options: &MarkdownDocsOptions) -> MarkdownDisplayFormat {
+    effective_display_format(options, options.index_format)
+}
+
+pub(super) fn effective_parameters_format(options: &MarkdownDocsOptions) -> MarkdownDisplayFormat {
+    effective_display_format(options, options.parameters_format)
+}
+
+pub(super) fn effective_members_format(
+    options: &MarkdownDocsOptions,
+    entry_kind: &str,
+    group_title: &str,
+) -> MarkdownDisplayFormat {
+    let format = match (entry_kind, group_title) {
+        ("class", "Properties" | "Static Properties") => options.class_properties_format,
+        ("interface", "Properties") => options.interface_properties_format,
+        ("type", "Properties") => options.type_alias_properties_format,
+        ("enum", "Enum Members" | "Members") | ("type", "Enum Members") => {
+            options.enum_members_format
+        }
+        _ => return MarkdownDisplayFormat::None,
+    };
+    effective_display_format(options, format)
+}
